Migrated modulus and phase calculations to CWTOutput and fixed errors in phase.
